Move ContactPicture to different package
