+++ app/core/src/main/java/com/fsck/k9/activity/misc/ContactPicture.java
@@ -1,11 +1,10 @@
-package com.fsck.k9.helper;
+package com.fsck.k9.activity.misc;
 
 import android.content.Context;
 import android.util.TypedValue;
 
 import com.fsck.k9.K9;
 import com.fsck.k9.core.R;
-import com.fsck.k9.activity.misc.ContactPictureLoader;
 
 public class ContactPicture {
