+++ app/k9mail/src/main/java/com/fsck/k9/directshare/K9ChooserTargetService.java
@@ -20,8 +20,8 @@ import android.support.annotation.Nullable;
 
 import com.fsck.k9.activity.MessageCompose;
 import com.fsck.k9.activity.compose.RecipientLoader;
+import com.fsck.k9.activity.misc.ContactPicture;
 import com.fsck.k9.activity.misc.ContactPictureLoader;
-import com.fsck.k9.helper.ContactPicture;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.view.RecipientSelectView.Recipient;
